+++ src/Intersection.h
@@ -7,27 +7,27 @@
 class Intersection {
 
 private:
-    float _t;
+    double _t;
     Shape _object;
 
 public:
     Intersection() : _t{0.0}, _object{Sphere()} { }
-    Intersection(const float t) : _t{t}, _object{Sphere()} { }
-    Intersection(const float t, const Shape s) : _t{t}, _object{s} {    }
+    Intersection(const double t) : _t{t}, _object{Sphere()} { }
+    Intersection(const double t, const Shape s) : _t{t}, _object{s} {    }
 
     bool operator==(const Intersection b) const {
         return (equal(_t, b.t()) && b.object().id()==_object.id() );
     }
 
     Shape object() const { return _object; }
-    float t() const { return _t; }
+    double t() const { return _t; }
 };
 
 /* HIT */
 std::shared_ptr<Intersection> hit(const vector<Intersection>& xs) {
 
     if (xs.size() == 0) return nullptr;
-    
+
     std::shared_ptr<Intersection> i=std::make_shared<Intersection>(10000);
 
     // find the element of xs with the lowest, non-negative t-value
